pyproject.toml
[tool.poetry]
name = "chess_bot"
version = "0.1.0"
description = "Discord bot that GIFs your chess games from lichess.org and chess.com!"
authors = ["Tomás Farías <tomas@tomasfarias.dev>"]
packages = [
    { include="chess_bot", from="." },
]

[tool.poetry.scripts]
chess-bot = "chess_bot.__main__:main"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
flake8 = "^3.8.3"
mypy = "^0.782"

[tool.black]
line-length = 110
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.pytest_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
